Модули по случайным данным теперь в одном общем модуле utils.js
