--- js/get-random-array-element.js
@@ -1,5 +1,0 @@
-import {getRandomInteger} from './get-random-integer.js';
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-export {getRandomArrayElement};
