--- js/get-random-integer.js
@@ -1,10 +1,0 @@
-function getRandomInteger(min, max) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  if (min >= 0 && max >= 0 && max > min && min !== String && max !== String) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-  return 'Исходные данные неверны';
-}
-
-export {getRandomInteger};
