+++ js/main.js
@@ -1,6 +1,4 @@
-import './get-random-integer.js';
-import './get-random-with-dot.js';
-import './get-random-array-element.js';
+import './utils.js';
 import {cardsOffers} from './create-new-offer.js';
 // eslint-disable-next-line no-console
 console.log(cardsOffers);
